+++ src/lein_monolith/plugin.clj
@@ -10,6 +10,32 @@
 
 ;; ## Profile Generation
 
+(def profile-config
+  "Configuration for inherited profiles. Structured as a vector of pairs to
+  maintain ordering. The ordering is significant as the info command consumes
+  this configuration directly, and providing deterministic output ordering is
+  desirable."
+  [[:monolith/inherited
+    {:inherit-key :inherit
+     :subproject-key :monolith/inherit}]
+
+   [:monolith/inherited-raw
+    {:raw? true
+     :inherit-key :inherit-raw
+     :subproject-key :monolith/inherit-raw}]
+
+   [:monolith/leaky
+    {:leaky? true
+     :inherit-key :inherit-leaky
+     :subproject-key :monolith/inherit-leaky}]
+
+   [:monolith/leaky-raw
+    {:leaky? true
+     :raw? true
+     :inherit-key :inherit-leaky-raw
+     :subproject-key :monolith/inherit-leaky-raw}]])
+
+
 (defn- subproject-dependencies
   "Given a map of internal projects, return a vector of dependency coordinates
   for the subprojects."
@@ -17,63 +43,58 @@
   (mapv #(vector (key %) (:version (val %))) subprojects))
 
 
-(defn- add-profile-paths
-  "Update a profile paths entry by adding the absolute paths from the given
-  project. Returns the updated profile."
-  [profile project k]
-  (update profile k into
-          (map (partial str (:root project) "/")
-               (get project k))))
+(defn- maybe-mark-leaky
+  "Add ^:leaky metadata to a profile if it is of the leaky type."
+  [profile {:keys [leaky?]}]
+  (if leaky?
+    (vary-meta profile assoc :leaky true)
+    profile))
 
 
-(defn merged-profile
-  "Constructs a profile map containing merged (re)source and test paths."
-  [subprojects]
-  (reduce-kv
-    (fn [profile _ project]
-      (-> profile
-          (add-profile-paths project :resource-paths)
-          (add-profile-paths project :source-paths)
-          (add-profile-paths project :test-paths)))
-    {:dependencies (subproject-dependencies subprojects)
-     :resource-paths []
-     :source-paths []
-     :test-paths []}
-    subprojects))
+(defn- choose-inheritance-source
+  "Choose either the initialized monolith or its raw representation for use when
+  building an inherited profile."
+  [monolith {:keys [raw?]}]
+  (if raw?
+    (:monolith/raw (meta monolith))
+    monolith))
 
 
 (defn- select-inherited-properties
-  "Constructs a profile map containing the given inherited properties from the
-  monolith project map."
-  [monolith base-properties setting]
-  (cond
-    ;; Don't inherit anything
-    (not setting)
-    nil
+  "Constructs a profile map containing the inherited properties from a parent
+  project map."
+  [monolith base-properties subproject subproject-key]
+  (let [default (boolean (:monolith/inherit subproject))
+        setting (subproject-key subproject default)]
+    (cond
+      ;; Don't inherit anything
+      (not setting)
+      nil
 
-    ;; Inherit the base properties specified in the parent.
-    (true? setting)
-    (select-keys monolith base-properties)
+      ;; Inherit the base properties specified in the parent.
+      (true? setting)
+      (select-keys monolith base-properties)
 
-    ;; Provide additional properties to inherit, or replace if metadata is set.
-    (vector? setting)
-    (select-keys monolith
-                 (if (:replace (meta setting))
-                   setting
-                   (distinct (concat base-properties setting))))
+      ;; Provide additional properties to inherit, or replace if metadata is set.
+      (vector? setting)
+      (select-keys monolith
+                   (if (:replace (meta setting))
+                     setting
+                     (distinct (concat base-properties setting))))
 
-    :else
-    (throw (ex-info (str "Unknown value type for monolith inherit setting: "
-                         (pr-str setting))
-                    {:inherit setting}))))
+      :else
+      (throw (ex-info (str "Unknown value type for monolith inherit setting: "
+                           (pr-str setting))
+                      {:inherit setting
+                       :subproject-key subproject-key})))))
 
 
 (defn- inherited-profile
   "Constructs a profile map containing the inherited properties from a parent
   project map."
-  [monolith inherit-key setting]
+  [monolith subproject {:keys [inherit-key subproject-key]}]
   (when-let [base-properties (get-in monolith [:monolith inherit-key])]
-    (when-let [profile (select-inherited-properties monolith base-properties setting)]
+    (when-let [profile (select-inherited-properties monolith base-properties subproject subproject-key)]
       (when (contains? profile :profiles)
         (lein/warn "WARN: nested profiles cannot be inherited; ignoring :profiles in monolith" inherit-key))
       (dissoc profile :profiles))))
@@ -82,16 +103,16 @@
 (defn build-inherited-profiles
   "Returns a map from profile keys to inherited profile maps."
   [monolith subproject]
-  (let [inherit-profile (inherited-profile
-                          monolith :inherit
-                          (:monolith/inherit subproject))
-        leaky-profile (inherited-profile
-                        monolith :inherit-leaky
-                        (:monolith/leaky subproject (boolean (:monolith/inherit subproject))))]
-    (cond-> nil
-      inherit-profile (assoc :monolith/inherited inherit-profile)
-      leaky-profile (assoc :monolith/leaky (vary-meta leaky-profile assoc :leaky true)))))
-
+  (reduce
+    (fn [acc [key config]]
+      (let [profile (some-> (choose-inheritance-source monolith config)
+                            (inherited-profile subproject config)
+                            (maybe-mark-leaky config))]
+        (if profile
+          (assoc acc key profile)
+          acc)))
+    nil
+    profile-config))
 
 
 ;; ## Profile Utilities
@@ -129,26 +150,29 @@
       (activate-profile profile-key)))
 
 
-
 ;; ## Plugin Middleware
 
 (defn middleware
   "Handles inherited properties in monolith subprojects by looking for the
   `:monolith/inherit` key."
-  [project]
-  (if (:monolith/inherit project)
-    ; Monolith subproject, add inherited profile.
-    (if (or (profile-active? project :monolith/inherited)
-            (profile-active? project :monolith/leaky))
-      ; Already activated, return project.
-      (do (lein/debug "One or both inherited profiles are already active!")
-          project)
-      ; Find monolith metaproject and generate profile.
-      (let [monolith (config/find-monolith! project)
-            profiles (build-inherited-profiles monolith project)]
-        (reduce-kv add-active-profile project profiles)))
-    ; Normal project, don't activate.
-    project))
+  ([project]
+   (middleware project nil))
+  ([project monolith]
+   (if (:monolith/inherit project)
+     ;; Monolith subproject, add inherited profile.
+     (if (some (fn this-profile-active?
+                 [entry]
+                 (profile-active? project (first entry)))
+               profile-config)
+       ;; Already activated, return project.
+       (do (lein/debug "One or more inherited profiles are already active!")
+           project)
+       ;; Find monolith metaproject and generate profile.
+       (let [monolith (or monolith (config/find-monolith! project))
+             profiles (build-inherited-profiles monolith project)]
+         (reduce-kv add-active-profile project profiles)))
+     ;; Normal project, don't activate.
+     project)))
 
 
 (defn add-middleware
@@ -159,3 +183,41 @@
     (if (some #{mw-sym} (:middleware project))
       project
       (update project :middleware (fnil conj []) mw-sym))))
+
+
+;; ## Merged Profiles (`with-all`) Creation
+
+
+(def ^:private path-keys
+  "Project map keys for (re)source and test paths."
+  #{:resource-paths :source-paths :test-paths})
+
+
+(defn- add-profile-paths
+  "Update a profile paths entry by adding the paths from the given project.
+  Returns the updated profile."
+  [project profile k]
+  (update profile k (fn combine-colls
+                      [coll]
+                      (-> coll
+                          set
+                          (into (get project k))
+                          (vary-meta assoc :replace true)))))
+
+
+(defn merged-profile
+  "Constructs a profile map containing merged (re)source and test paths."
+  [monolith subprojects]
+  (let [profile
+        (reduce-kv
+          (fn [profile _project-name subproject]
+            (let [with-inherited-profiles (middleware subproject monolith)
+                  project (project/absolutize-paths with-inherited-profiles)]
+              (reduce (partial add-profile-paths project)
+                      profile
+                      path-keys)))
+          (select-keys monolith path-keys)
+          subprojects)]
+    (as-> profile v
+          (reduce (fn sort-paths [acc k] (update acc k sort)) v path-keys)
+          (assoc v :dependencies (subproject-dependencies subprojects)))))